requirements.txt
Django==2.2.4
django-filter==2.2.0
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
mysqlclient==1.4.2
whitenoise==4.1.3
python-dotenv